Replace all upgrade-related abstractions under tendermint host
