+++ crates/ibc/src/applications/transfer/context.rs
@@ -19,7 +19,7 @@ use crate::core::ics04_channel::context::{
 use crate::core::ics04_channel::packet::{Acknowledgement, Packet};
 use crate::core::ics04_channel::Version;
 use crate::core::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
-use crate::core::ics26_router::ModuleExtras;
+use crate::core::router::ModuleExtras;
 use crate::core::ContextError;
 use crate::signer::Signer;
