--- crates/ibc/src/core/context/mod.rs
@@ -1,7 +1,0 @@
-mod error;
-mod val_exec;
-
-pub use error::ContextError;
-pub use val_exec::{ExecutionContext, ValidationContext};
-
-pub mod upgrade_client;
