+++ crates/ibc/src/core/handler.rs
@@ -1,3 +1,4 @@
+use super::context::RouterError;
 use super::ics02_client::handler::{create_client, update_client, upgrade_client};
 use super::ics02_client::msgs::{ClientMsg, MsgUpdateOrMisbehaviour};
 use super::ics03_connection::handler::{
@@ -27,7 +28,6 @@ use super::ics04_channel::handler::timeout::{
     timeout_packet_execute, timeout_packet_validate, TimeoutMsgType,
 };
 use super::ics04_channel::msgs::{ChannelMsg, PacketMsg};
-use super::ics26_router::RouterError;
 use super::ContextError;
 use super::{msgs::MsgEnvelope, ExecutionContext, ValidationContext};
 
@@ -226,8 +226,8 @@ mod tests {
     use crate::core::ics23_commitment::commitment::CommitmentPrefix;
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
     use crate::core::ics24_host::path::CommitmentPath;
-    use crate::core::ics26_router::ModuleId;
     use crate::core::msgs::MsgEnvelope;
+    use crate::core::router::ModuleId;
     use crate::core::timestamp::Timestamp;
     use crate::mock::client_state::MockClientState;
     use crate::mock::consensus_state::MockConsensusState;
