+++ crates/ibc/src/core/ics02_client/mod.rs
@@ -9,4 +9,3 @@ pub mod handler;
 pub mod header;
 pub mod height;
 pub mod msgs;
-pub mod proposal;
