+++ crates/ibc/src/core/ics04_channel/handler/recv_packet.rs
@@ -14,7 +14,7 @@ use crate::core::ics24_host::path::Path;
 use crate::core::ics24_host::path::{
     AckPath, ChannelEndPath, ClientConsensusStatePath, CommitmentPath, ReceiptPath, SeqRecvPath,
 };
-use crate::core::ics26_router::ModuleId;
+use crate::core::router::ModuleId;
 use crate::core::timestamp::Expiry;
 use crate::core::{ContextError, ExecutionContext, ValidationContext};
