--- crates/ibc/src/core/ics26_router/mod.rs
@@ -1,7 +1,0 @@
-mod error;
-mod module;
-mod router;
-
-pub use error::RouterError;
-pub use module::{Module, ModuleExtras, ModuleId};
-pub use router::Router;
