+++ crates/ibc/src/core/msgs.rs
@@ -2,6 +2,7 @@ use crate::prelude::*;
 
 use ibc_proto::google::protobuf::Any;
 
+use crate::core::context::RouterError;
 use crate::core::ics02_client::msgs::{
     create_client, misbehaviour, update_client, upgrade_client, ClientMsg,
 };
@@ -12,7 +13,6 @@ use crate::core::ics04_channel::msgs::{
     acknowledgement, chan_close_confirm, chan_close_init, chan_open_ack, chan_open_confirm,
     chan_open_init, chan_open_try, recv_packet, timeout, timeout_on_close, ChannelMsg, PacketMsg,
 };
-use crate::core::ics26_router::RouterError;
 use ibc_proto::protobuf::Protobuf;
 
 /// Trait to be implemented by all IBC messages
