+++ crates/ibc/src/hosts/tendermint/upgrade_proposal/context.rs
@@ -2,12 +2,11 @@ use alloc::boxed::Box;
 
 use crate::core::ics02_client::client_state::ClientState;
 use crate::core::ics02_client::consensus_state::ConsensusState;
-use crate::core::ics02_client::proposal::Plan;
 use crate::core::ics24_host::path::UpgradeClientPath;
+use crate::core::{ContextError, ExecutionContext, ValidationContext};
 use crate::Height;
 
-use super::error::ContextError;
-use super::{ExecutionContext, ValidationContext};
+use super::proposal::Plan;
 
 /// Extends `ValidationContext` capability by providing required methods to
 /// handle an upgrade client proposal.
