+++ crates/ibc/src/hosts/tendermint/upgrade_proposal/mod.rs
@@ -1,2 +1,4 @@
+pub mod context;
 pub mod handler;
 pub mod helper;
+pub mod proposal;
