+++ crates/ibc/src/test_utils.rs
@@ -5,7 +5,7 @@ use crate::core::ics04_channel::error::{ChannelError, PacketError};
 use crate::core::ics04_channel::packet::{Acknowledgement, Packet};
 use crate::core::ics04_channel::Version;
 use crate::core::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
-use crate::core::ics26_router::{Module, ModuleExtras};
+use crate::core::router::{Module, ModuleExtras};
 use crate::prelude::*;
 use crate::signer::Signer;
